% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ames_HCD}
\alias{ames_HCD}
\title{Historical numbers of revertant colonies in the Ames test (OECD 471)}
\format{
A \code{data.frame} with 2 rows and 10 columns:
 \describe{
  \item{rev_col}{no. of revertant colonies}
  \item{no_dish}{no. of petri dishes in the control group}
}
}
\usage{
ames_HCD
}
\description{
This data set contains artificial historical control data that was sampled in
order to mimic the number of revertant colonies based on two or three petri dishes.
}
\keyword{datasets}
