% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_bin_pi.R
\name{beta_bin_pi}
\alias{beta_bin_pi}
\title{beta_bin_pi}
\usage{
beta_bin_pi(
  histdat,
  newdat = NULL,
  newsize = NULL,
  alternative = "both",
  alpha = 0.05,
  nboot = 10000,
  lambda_min = 0.01,
  lambda_max = 10,
  traceplot = TRUE,
  n_bisec = 30
)
}
\arguments{
\item{histdat}{a data frame containing the historical data}

\item{newdat}{a data frame containing the actual data}

\item{newsize}{a vector containing new cluster sizes}

\item{alternative}{either "both", "upper" or "lower"}

\item{n_bisec}{}
}
\value{
a data frame with prediction intervals
}
\description{
beta_bin_pi calculates prediction intervals for beta binomial distributed one way design
}
