% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_predint.R
\name{boot_predint}
\alias{boot_predint}
\title{Bootstrap new data from uncalibrated prediction intervals}
\usage{
boot_predint(pred_int, nboot)
}
\arguments{
\item{pred_int}{simple prediction interval}

\item{nboot}{number of bootstraps}
}
\value{
\code{boot_predint} returns an object of class \code{c("predint", "bootstrap")}
which is a list with two entries: One for bootstrapped historical observations
and one for bootstrapped future observations.
}
\description{
\code{boot_predint} is a helper function to bootstrap new data from the simple
uncalibrated prediction intervals implemented in predint
}
\examples{
test_pi <- qp_pi(histoffset=c(1:9), newoffset=c(3), lambda=10, phi=3, q=qnorm(1-0.05/2), alternative="both")

test_boot <- boot_predint(pred_int = test_pi, nboot=5)
test_boot

}
