% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{mortality_HCD}
\alias{mortality_HCD}
\title{Historical mortality of male B6C3F1-mice}
\format{
A \code{data.frame} with 2 rows and 10 columns:
 \describe{
  \item{dead}{no. of dead mice}
  \item{alive}{no. of living mice}
}
}
\usage{
mortality_HCD
}
\description{
This data set contains historical control data about the mortality of male B6C3F1-mice
obtained in long term carcinogenicity studies at the National Toxicology Program
presented in NTP Historical Control Reports from 2013 to 2016.
It was used in Menssen and Schaarschmidt 2019 as a real life example.
}
\references{
Menssen and Schaarschmidt (2019): Prediction intervals for overdispersed binomial
data with application to historical controls. Statistics in Medicine.
\doi{10.1002/sim.8124} \cr
NTP Historical Control Reports: \url{https://ntp.niehs.nih.gov/data/controls}
}
\keyword{datasets}
