% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_bin_pi_rho_estimation.R
\name{pi_rho_est}
\alias{pi_rho_est}
\title{Estimation of \eqn{\pi} and \eqn{\rho} of beta-binomial data according
to Lui et al. 2000.}
\usage{
pi_rho_est(dat)
}
\arguments{
\item{dat}{a data frame with two columns (success and failure)}
}
\value{
a vector containing estimates for \eqn{\pi} and \eqn{\rho}
}
\description{
pi_rho_est estimates the overall binomial proportion \eqn{\pi} and the intra
class correlation \eqn{\rho} of data that is assumed to follow the beta-binomial
distribution
}
\examples{
# Estimates for the bb_dat1
pi_rho_est(bb_dat1)

}
\references{
Lui, K.-J., Mayer, J.A. and Eckhardt, L: Confidence intervals for the risk ratio
under cluster sampling based on the beta-binomial model. Statistics in Medicine.2000;19:2933-2942.
https://doi.org/10.1002/1097-0258(20001115)19:21<2933::AID-SIM591>3.0.CO;2-Q
}
