% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.predint.R
\name{plot.predint}
\alias{plot.predint}
\title{Plots of prediction intervals}
\usage{
\method{plot}{predint}(x, size = 4, width = 0.05, alpha = 0.5)
}
\arguments{
\item{scale}{}
}
\value{

}
\description{
Plots of prediction intervals
}
