% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.predint.R
\name{print.predint}
\alias{print.predint}
\title{Print function for objects of class \code{predint}}
\usage{
\method{print}{predint}(x)
}
\arguments{
\item{x}{an object of class \code{predint}}
}
\value{
prints output to the console
}
\description{
Print function for objects of class \code{predint}
}
\examples{
# see examples for the calculation of prediction intervals
}
