% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{qb_dat1}
\alias{qb_dat1}
\title{Quasi-binomial data (example 1)}
\format{
A \code{data.frame} with 3 rows and 2 columns:
\describe{
  \item{succ}{numbers of success}
  \item{fail}{numbers of failures}
}
}
\usage{
qb_dat1
}
\description{
This data set contains sampled quasi-binomial data from 10 clusters
each of size 50. The data set was sampled with \code{rqbinom(n=10, size=50, prob=0.1, phi=3)}.
}
\keyword{datasets}
