% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{qp_dat1}
\alias{qp_dat1}
\title{Quasi-Poisson data (example 1)}
\format{
An integer vector with ten entries containing quasi-Poisson data
}
\usage{
qp_dat1
}
\description{
This data set contains sampled quasi-poisson data for 10 clusters.
The data set was sampled with \code{rqpois(n=10, lambda=50, phi=3)}.
}
\examples{
# Upper prediction limit for m=3 future observations
quasi_pois_pi(histdat=data.frame(qp_dat1), m=3, alternative="upper", nboot=100)

# Please note that nboot was set to 100 in order to decrease computing time
# of the example. For a valid analysis set nboot=10000.

}
\keyword{datasets}
