% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_pi.R
\name{qp_pi}
\alias{qp_pi}
\title{Simple uncalibrated prediction intervals for quasi-Poisson data}
\usage{
qp_pi(
  newoffset,
  histoffset,
  lambda,
  phi,
  q = qnorm(1 - 0.05/2),
  alternative = "both"
)
}
\arguments{
\item{newoffset}{number of experimental units in the future clusters.}

\item{histoffset}{number of experimental units in the historical clusters.}

\item{lambda}{overall Poisson mean.}

\item{phi}{dispersion parameter}

\item{q}{quantile used for interval calculation}

\item{alternative}{either "both", "upper" or "lower".
\code{alternative} specifies, if a prediction interval or
an upper or a lower prediction limit should be computed.}
}
\value{
\code{qp_pi} returns an object of class \code{c("predint", "qpUncalib")}.
}
\description{
\code{qp_pi()} is a helper fuction that is called in \code{quasi_pois_pi()}. It
calculates simple uncalibrated prediction intervals for Poisson
data with constant overdispersion (quasi-Poisson).
}
\details{
This function returns a simple uncalibrated prediction interval
\deqn{[l,u] = n^*_m \hat{\lambda} \pm q \sqrt{n^*_m \hat{\phi} \hat{\lambda} +
n^{*2}_m \hat{\phi} \hat{\lambda} \frac{1}{\sum_h n_h}}}

with \eqn{n^*_m} as the number of experimental units in the future clusters,
\eqn{\hat{\lambda}} as the estimate for the Poisson mean obtained from the
historical data, \eqn{\hat{\phi}} as the estimate for the dispersion parameter
and \eqn{n_h} as the number of experimental units per historical cluster. \cr

The use of this uncalibrated prediction interval is not recommended for practical
application.  \cr
}
\examples{
# Prediction interval
qp_pi(newoffset=c(3), lambda=3, ph=3, histoffset=1:9, q=qnorm(1-0.05/2))

# Upper prediction border
qp_pi(newoffset=c(1), lambda=3, ph=3, histoffset=1:9, q=qnorm(1-0.05), alternative="upper")


}
