% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_bin_sampling.R
\name{rbbinom}
\alias{rbbinom}
\title{Sampling of beta binomial data}
\usage{
rbbinom(n, size, prob, rho)
}
\arguments{
\item{n}{defines the number of clusters (\eqn{i})}

\item{size}{integer vector defining the cluster sizes (\eqn{n_i})}

\item{prob}{overall binomial proportion (\eqn{\pi})}

\item{rho}{intra class correlation (\eqn{\rho})}
}
\value{
a data frame with two columns (succ, fail)
}
\description{
rbbinom samples beta binomial data according to Menssen and Schaarschmidt (2019)
}
\details{
The sampling is done such that
\deqn{x_i \sim Bin(n_i, \pi_i) and \bi_i \sim Beta(a,b)}
with \eqn{i=1..I} clusters where \eqn{E(\pi_i)=\pi=a/(a+b)} and
\eqn{E(x_i)=n_i \pi}. Hence, the beta binomial variance is
\deqn{var(x_i)=n_i \pi (1-\pi) (1+ (n_i-1) \rho)}
with \eqn{\rho} as the intra class correlation
\deqn{\rho = \frac{1}{1+a+b}}
Please note, that \eqn{1+ (n_i-1) \rho} is a constant if all cluster sizes are
the same and hence, in this special case, also the quasi binomial assumption is
fulfilled.
}
\examples{
# Sampling of example data
set.seed(234)
bb_dat1 <- rbbinom(n=10, size=50, prob=0.1, rho=0.06)
bb_dat1


set.seed(234)
bb_dat2 <- rbbinom(n=3, size=c(40, 50, 60), prob=0.1, rho=0.06)
bb_dat2


}
\references{
Menssen M, Schaarschmidt F.: Prediction intervals for overdispersed binomial data
with application to historical controls. Statistics in Medicine. 2019;38:2652-2663.
https://doi.org/10.1002/sim.8124
}
