% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnbinom.R
\name{rnbinom}
\alias{rnbinom}
\title{Sampling of negative-binomial data}
\usage{
rnbinom(n, lambda, kappa, offset = NULL)
}
\arguments{
\item{n}{defines the number of clusters (\eqn{I})}

\item{lambda}{defines the overall Poisson mean (\eqn{\lambda})}

\item{kappa}{dispersion parameter (\eqn{\kappa})}

\item{offset}{defines the number of experimental units per cluster (\eqn{n_i})}
}
\value{

}
\description{
\code{rnbinom()} samples negative-binomial distributed data considering offsets.
The following description of the sampling process is based on the parametrization
used by Gsteiger et al. 2013.
}
\details{
In order to sample \eqn{i=1, ... I} observations,
the parameters of the gamma distribution are set to \eqn{a=1/\kappa} and
\eqn{b_i=1/(\kappa n_i \lambda)}.
Then, the Poisson means for each cluster are sampled from the gamma distribution
\deqn{\lambda_i \sim Gamma(a_, b_i)}
and the observations per cluster are sampled to be
\deqn{y_i \sim Pois(\lambda_i).}
Please note, that the negative-binomial distribution is not in contradiction with the
quasi-Poisson assumption, if the data structure is defined by the number
of clusters only (which is the case here) and the offsets are all the same
\eqn{n_h = n_{h´} = n}.
}
\examples{

# No offsets: Same results as rqpois(n=10, lambda=50, phi=3)
set.seed(123)
rnbinom(n=10, lambda=50, kappa=0.04)

# Different offsets
set.seed(123)
rnbinom(n=3, lambda=50, kappa=0.04, offset=c(3,1,0.5))


}
