% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{bb_dat2}
\alias{bb_dat2}
\title{Beta-binomial data (example 2)}
\format{
A data frame with 3 rows and 2 columns:
\describe{
  \item{succ}{number of succes}
  \item{fail}{number of failure}
}
}
\usage{
bb_dat2
}
\description{
This data set contains sampled beta-binomial data from 3 clusters with
different size, each.
}
\examples{
# Sampling of bb_dat1
set.seed(234)
bb_dat1 <- rbbinom(n=10, size=50, prob=0.1, rho=0.06)
bb_dat1

# Sampling of bb_dat2
set.seed(234)
bb_dat2 <- rbbinom(n=3, size=c(40, 50, 60), prob=0.1, rho=0.06)
bb_dat2

# Prediction interval using bb_dat2 as future data
beta_bin_pi(histdat=bb_dat1, newdat=bb_dat2)

# Upper prediction bound for m=3 future number of successes
# that are based on cluster sizes 40, 50, 60 respectively
beta_bin_pi(histdat=bb_dat1, newsize=c(40, 50, 60), alternative="upper")

}
\keyword{datasets}
