% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_bin_pi.R
\name{beta_bin_pi}
\alias{beta_bin_pi}
\title{Prediction intervals for beta binomial data}
\usage{
beta_bin_pi(
  histdat,
  newdat = NULL,
  newsize = NULL,
  alternative = "both",
  alpha = 0.05,
  nboot = 10000,
  lambda_min = 0.01,
  lambda_max = 10,
  traceplot = TRUE,
  n_bisec = 30
)
}
\arguments{
\item{histdat}{a data frame with two columns (success and failure) containing the historical data}

\item{newdat}{a data frame with two columns (success and failure) containing the future data}

\item{newsize}{a vector containing new cluster sizes}

\item{alternative}{either "both", "upper" or "lower" specifying if a prediction interval or
an upper or a lower prediction limit should be computed}

\item{alpha}{defines the level of confidence (1-alpha)}

\item{nboot}{number of bootstraps}

\item{lambda_min}{lower start value for bisection}

\item{lambda_max}{upper start value for bisection}

\item{traceplot}{plot for visualization of the bisection process}

\item{n_bisec}{maximal number of bisection steps}
}
\value{
If newdat is specified: A data frame that contains the future data,
 the historical mean (hist_mean), the calibrated coefficient (quant_calib),
 the prediction error (pred_se), the prediction interval (lower and upper)
 and a statement if the prediction interval covers the future observation (cover).

 If m is specified: A data frame that contains the number of future observations (m)
 the historical mean (hist_mean), the calibrated coefficient (quant_calib),
 the prediction error (pred_se) and the prediction interval (lower and upper).

 If alternative is set to "lower": Lower prediction bounds are computed instead
 of a prediction interval.

 If alternative is set to "upper": Upper prediction bounds are computed instead
 of a prediction interval.
}
\description{
beta_bin_pi calculates bootstrap calibrated prediction intervals for beta
binomial data
}
\details{
This function returns a bootstrap calibrated prediction interval
\deqn{[l,u]_m = \hat{y}_m \pm q \sqrt{var(\hat{y}_m - y_m)}}
with \eqn{\hat{y}}_m as the predicted future number of successes for \eqn{m=1,...,M} future clusters,
\eqn{y}_m as the observed future number of successes, \eqn{\sqrt{var(\hat{y}_m - y_m)}}
as the prediction error and \eqn{q} as the bootstrap calibrated coefficient that
approximates a multivariate normal distribution. Please note that
the predicted future number of successes is based on the future cluster size
and the success probability estimated from the historical data
\eqn{\hat{y}=\pi^{hist} n^{fut}}.

The whole calibration process is based on a bisection algorithm that is similar
to the one described in Menssen and Schaarschmidt 2019. If traceplot=TRUE, a graphical
overview about the bisection process is given.
}
\examples{
# Historical data
bb_dat1

# Future data
bb_dat2

# Prediction interval using bb_dat2 as future data
beta_bin_pi(histdat=bb_dat1, newdat=bb_dat2)

# Upper prediction bound for m=3 future number of successes
# that are based on cluster sizes 40, 50, 60 respectively
beta_bin_pi(histdat=bb_dat1, newsize=c(40, 50, 60), alternative="upper")


}
\references{
Menssen M, Schaarschmidt F.: Prediction intervals for overdispersed binomial data
with application to historical controls. Statistics in Medicine. 2019;38:2652-2663.
https://doi.org/10.1002/sim.8124
}
