% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{c2_dat3}
\alias{c2_dat3}
\title{Cross-classified data (example 3)}
\format{
A \code{data.frame} with 8 rows and 3 columns:
\describe{
  \item{y_ijk}{observations}
  \item{a}{treatment a}
  \item{b}{treatment b}
}
}
\usage{
c2_dat3
}
\description{
c2_dat3 contains data that was sampled from a balanced cross-classified design.
This data set is used in order to demonstrate the functionality of the \code{lmer_pi_...()} functions.
}
\keyword{datasets}
