% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_pi_futmat.R
\name{lmer_pi_futmat}
\alias{lmer_pi_futmat}
\title{Prediction intervals for future observations based on linear random effects models}
\usage{
lmer_pi_futmat(
  model,
  newdat = NULL,
  futmat_list = NULL,
  alternative = "both",
  alpha = 0.05,
  nboot = 10000,
  lambda_min = 0.01,
  lambda_max = 10,
  traceplot = TRUE,
  n_bisec = 30
)
}
\arguments{
\item{model}{a random effects model of class lmerMod}

\item{newdat}{either 1 or a \code{data.frame} with the same column names as the historical data
on which \code{model} depends}

\item{alternative}{either "both", "upper" or "lower". \code{alternative} specifies
if a prediction interval or an upper or a lower prediction limit should be computed}

\item{alpha}{defines the level of confidence (1-\code{alpha})}

\item{nboot}{number of bootstraps}

\item{lambda_min}{lower start value for bisection}

\item{lambda_max}{upper start value for bisection}

\item{traceplot}{plot for visualization of the bisection process}

\item{n_bisec}{maximal number of bisection steps}
}
\value{
If \code{newdat} is a  \code{data.frame}: A \code{data.frame} that contains the future data,
 the historical mean (hist_mean), the calibrated coefficient (quant_calib),
 the prediction standard error (pred_se), the prediction interval (lower and upper)
 and a statement if the prediction interval covers the future observation (cover).

 If \code{newdat=1}: A \code{data.frame} that contains a statement that m=1,
 the historical mean (hist_mean), the calibrated coefficient (quant_calib),
 the prediction standard error (pred_se) and the prediction interval (lower and upper).

 If \code{alternative} is set to "lower": Lower prediction limits are computed instead
 of a prediction interval.

 If \code{alternative} is set to "upper": Upper prediction limits are computed instead
 of a prediction interval.

 If \code{traceplot=TRUE}, a graphical overview about the bisection process is given.
}
\description{
lmer_pi_futmat calculates a bootstrap calibrated prediction interval for one or more
future observation(s) based on linear random effects models. With this approach,
the sampling structure of the future data is taken into account (see below).
}
\details{
This function returns a bootstrap calibrated prediction interval
\deqn{[l,u] = \hat{y} \pm q \sqrt{\hat{var}(\hat{y} - y)}}
with \eqn{\hat{y}} as the predicted future observation,
\eqn{y} as the observed future observations, \eqn{\sqrt{\hat{var}(\hat{y} - y)}}
as the prediction standard error and \eqn{q} as the bootstrap calibrated coefficient that
approximates a multivariate t-distribution. \cr
Please note that this function relies on linear random effects models that are
fitted with lmer() from the lme4 package.Random effects have to be specified as
\code{(1|random_effect)}.\cr

Be aware that the bootstrapped future observations used for the calibration
process mimic the structure of the data set provided via \code{newdat}. The
data sampling is based on the design matrix that can be obtained if \code{newdat}
and the model formula are provided to \code{lme4::lFormula()}. Hence, each random
factor that is part of the initial model must have at least two replicates in
\code{newdat}.

This function is similar to the PI given in Menssen and Schaarschmidt 2021
section 3.2.4 except that the bootstrap calibration values are drawn from
bootstrap samples that mimic the future data as described above.
}
\examples{

# loading lme4
library(lme4)

# Fitting a random effects model based on c2_dat_1
fit <- lmer(y_ijk~(1|a)+(1|b)+(1|a:b), c2_dat1)
summary(fit)

# Prediction interval using c2_dat2 as future data
lmer_pi_futmat(model=fit, newdat=c2_dat2, alternative="both", nboot=100)

# Upper prediction limit for m=1 future observations
lmer_pi_futmat(model=fit, newdat=1, alternative="upper", nboot=100)

# Please note that nboot was set to 100 in order to increase computing time
# of the example. For a valid analysis set nboot=10000.

}
\references{
Menssen, M., Schaarschmidt, F.: Prediction intervals for all of M
future observations based on linear random effects models. Statistica Neerlandica.
\url{https://doi.org/10.1111/stan.12260}
}
