% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.predint.R
\name{print.predint}
\alias{print.predint}
\title{Print objects of class \code{predint}}
\usage{
\method{print}{predint}(x, ...)
}
\arguments{
\item{x}{an object of class \code{predint}}

\item{...}{additional arguments passed over to \code{base::cbind()} and \code{base::data.frame()}}
}
\value{
prints output to the console
}
\description{
Print objects of class \code{predint}
}
