% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{qb_dat2}
\alias{qb_dat2}
\title{Quasi-binomial data (example 2)}
\format{
A data frame with 3 rows and 2 variables:
\describe{
  \item{succ}{number of succes}
  \item{fail}{number of failure}
}
}
\usage{
qb_dat2
}
\description{
This data set contains sampled quasi binomial data from 3 clusters with
different size
}
\examples{
# Sampling of qb_dat1
set.seed(456)
qb_dat1 <- rqbinom(n=10, size=50, prob=0.1, phi=3)
qb_dat1

# Sampling of qb_dat2
set.seed(456)
qb_dat2 <- rqbinom(n=3, size=c(40, 50, 60), prob=0.1, phi=3)
qb_dat2

# Prediction interval using qb_dat2 as future data
quasi_bin_pi(histdat=qb_dat1, newdat=qb_dat2)

# Upper prediction bound for m=3 future observations
# that are based on cluster sizes 40, 50, 60 respectively
quasi_bin_pi(histdat=qb_dat1, newsize=c(40, 50, 60), alternative="upper")

}
\keyword{datasets}
