% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{qp_dat1}
\alias{qp_dat1}
\title{Quasi-Poisson data (example 1)}
\format{
A data.frame with two columns
\describe{
  \item{y}{numbers of eventzs}
  \item{offset}{size of experimental units}
}
}
\usage{
qp_dat1
}
\description{
This data set contains sampled quasi-Poisson data for 10 clusters.
The data set was sampled with \code{rqpois(n=10, lambda=50, phi=3)}.
}
\keyword{datasets}
