% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{qp_dat2}
\alias{qp_dat2}
\title{Quasi-poisson data (example 2)}
\format{
A vector containing quasi poisson data:
}
\usage{
qp_dat2
}
\description{
This data set contains sampled quasi-poisson data for 3 clusters
}
\examples{
set.seed(123)
qp_dat1 <- rqpois(n=10, lambda=50, phi=3)
qp_dat1

set.seed(123)
qp_dat2 <- rqpois(n=3, lambda=50, phi=3)
qp_dat2

# Prediction interval using bb_dat2 as future data
quasi_pois_pi(histdat=data.frame(qp_dat1),
              newdat=data.frame(qp_dat2))

# Upper prediction bound for m=3 future observations
quasi_pois_pi(histdat=data.frame(qp_dat1),
              m=3,
              alternative="upper")

}
\keyword{datasets}
