% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi_pois_pi.R
\name{quasi_pois_pi}
\alias{quasi_pois_pi}
\title{Prediction intervals for quasi-Poisson data}
\usage{
quasi_pois_pi(
  histdat,
  newdat = NULL,
  newoffset = NULL,
  alternative = "both",
  alpha = 0.05,
  nboot = 10000,
  delta_min = 0.01,
  delta_max = 10,
  tolerance = 0.001,
  traceplot = TRUE,
  n_bisec = 30
)
}
\arguments{
\item{histdat}{a \code{data.frame} with two columns. The first has to contain
the historical observations. The second has to contain the number of experimental
units per study (offsets).}

\item{newdat}{a \code{data.frame} with two columns. The first has to contain
the future observations. The second has to contain the number of experimental
units per study (offsets).}

\item{newoffset}{vector with future number of experimental units per historical
study.}

\item{alternative}{either "both", "upper" or "lower".
\code{alternative} specifies if a prediction interval or
an upper or a lower prediction limit should be computed}

\item{alpha}{defines the level of confidence (\eqn{1-\alpha})}

\item{nboot}{number of bootstraps}

\item{delta_min}{lower start value for bisection}

\item{delta_max}{upper start value for bisection}

\item{tolerance}{tolerance for the coverage probability in the bisection}

\item{traceplot}{if \code{TRUE}: Plot for visualization of the bisection process}

\item{n_bisec}{maximal number of bisection steps}
}
\value{
\code{qp_pi} returns an object of class \code{c("predint", "quasiPoissonPI")}
with prediction intervals or limits in the first entry (\code{$prediction}).
}
\description{
quasi_pois_pi calculates bootstrap calibrated prediction intervals for Poisson
data with constant overdispersion (quasi-Poisson).
}
\details{
This function returns a bootstrap-calibrated prediction interval

\deqn{[l,u] = n^*_m \hat{\lambda} \pm q^{calib} \sqrt{n^*_m \hat{\phi} \hat{\lambda} +
n^{*2}_m \hat{\phi} \hat{\lambda} \frac{1}{\sum_h n_h}}}

with \eqn{n^*_m} as the number of experimental units in the future clusters,
\eqn{\hat{\lambda}} as the estimate for the Poisson mean obtained from the
historical data, \eqn{q^{calib}} as the bootstrap-calibrated coefficient,
\eqn{\hat{\phi}} as the estimate for the dispersion parameter
and \eqn{n_h} as the number of experimental units per historical cluster. \cr
}
\examples{
#' # Historical data
qp_dat1

# Future data
qp_dat2

# Prediction interval using bb_dat2 as future data
quasi_pois_pi(histdat=data.frame(qp_dat1), newdat=data.frame(qp_dat2), nboot=100)

# Upper prediction bound for m=3 future observations
quasi_pois_pi(histdat=data.frame(qp_dat1), newoffset=c(1,1,1), alternative="upper", nboot=100)

# Please note that nboot was set to 100 in order to decrease computing time
# of the example. For a valid analysis set nboot=10000.

}
