% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi_pois_pi.R
\name{quasi_pois_pi}
\alias{quasi_pois_pi}
\title{Prediction intervals for quasi-poisson data}
\usage{
quasi_pois_pi(
  histdat,
  newdat = NULL,
  m = NULL,
  alternative = "both",
  alpha = 0.05,
  nboot = 10000,
  lambda_min = 0.01,
  lambda_max = 10,
  traceplot = TRUE,
  n_bisec = 30
)
}
\arguments{
\item{histdat}{a data frame with one column containing the historical data}

\item{newdat}{a data frame with one column containing the future data}

\item{m}{number of future clusters}

\item{alternative}{either "both", "upper" or "lower". alternative specifies if a prediction interval or
an upper or a lower prediction limit should be computed}

\item{alpha}{defines the level of confidence (1-alpha)}

\item{nboot}{number of bootstraps}

\item{lambda_min}{lower start value for bisection}

\item{lambda_max}{upper start value for bisection}

\item{traceplot}{plot for visualization of the bisection process}

\item{n_bisec}{maximal number of bisection steps}
}
\value{
If newdat is specified: A data frame that contains the future data,
 the historical mean (hist_mean), the calibrated coefficient (quant_calib),
 the prediction error (pred_se), the prediction interval (lower and upper)
 and a statement if the prediction interval covers the future observation (cover).

 If m is specified: A data frame that contains the number of future observations (m)
 the historical mean (hist_mean), the calibrated coefficient (quant_calib),
 the prediction error (pred_se) and the prediction interval (lower and upper).

 If alternative is set to "lower": Lower prediction bounds are computed instead
 of a prediction interval.

 If alternative is set to "upper": Upper prediction bounds are computed instead
 of a prediction interval.
}
\description{
quasi_pois_pi calculates bootstrap calibrated prediction intervals for poisson
data with constant overdispersion (quasi-poisson).
}
\details{
This function returns a bootstrap calibrated prediction interval
\deqn{[l,u] = \hat{y} \pm q \sqrt{var(\hat{y} - y)}}
with \eqn{\hat{y}} as the predicted future observation,
\eqn{y} as the observed future observations, \eqn{\sqrt{var(\hat{y} - y)}}
as the prediction error and \eqn{q} as the bootstrap calibrated coefficient that
approximates a multivariate normal distribution.

The whole calibration process is based on a bisection algorithm that is similar
to the one described in Menssen and Schaarschmidt 2019. If traceplot=TRUE, a graphical
overview about the bisection process is given.
}
\examples{
#' # Historical data
qp_dat1

# Future data
qp_dat2

# Prediction interval using bb_dat2 as future data
quasi_pois_pi(histdat=data.frame(qp_dat1),
            newdat=data.frame(qp_dat2))

# Upper prediction bound for m=3 future observations
quasi_pois_pi(histdat=data.frame(qp_dat1),
              m=3,
              alternative="upper")



}
\references{
Menssen M, Schaarschmidt F.: Prediction intervals for overdispersed binomial data
with application to historical controls. Statistics in Medicine. 2019;38:2652-2663.
https://doi.org/10.1002/sim.8124
}
