% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi_pois_sampling.R
\name{rqpois}
\alias{rqpois}
\title{Sampling of overdispersed poisson data with constant overdispersion}
\usage{
rqpois(n, lambda, phi)
}
\arguments{
\item{n}{defines the number of clusters (\eqn{i})}

\item{lambda}{defines the overall poisson mean (\eqn{\lambda})}

\item{phi}{dispersion parameter (\eqn{\Phi})}
}
\value{

}
\description{
rqpois samples overdispersed poisson data with constant overdispersion from
the negative binomial distribution such that the quasi poisson assumption is fulfilled.
}
\details{
It is assumed that the dispersion parameter (\eqn{\Phi})
is constant for all \eqn{i=1, ... I} clusters, such that the variance becomes
\deqn{var(y_i)=\Phi \lambda_i.}
}
\references{
Gsteiger, S., Neuenschwander, B., Mercier, F. and Schmidli, H. (2013):
Using historical control information for the design and analysis of clinical
trials with overdispersed count data. Statist. Med., 32: 3609-3622. https://doi.org/10.1002/sim.5851
}
